Refactor model param vector to remove fixed params
Note that this is an untested commit.

We were attempting to update the model parameters vector with a mode
that was too short. The reason is that the new parameters vector
includes the "super fixed" parameters that were previously represented
as floats. Here, we change those to Param types and collect them
separately.

It would certainly be preferable to collect these "super fixed"
parameters normally in the model parameters vector. That might come with
a tiny computational cost but probably worth it. The hessian then would
need to be recalculated.