Savepaths created upon construction of model object if they do not already exist
This facilitates the organization of saved files for when we have additional types of models. The save directory structure looks as follows:

/DSGE
     /docs
	...
     /save
	 /modelspec           #e.g. m990/
	     /input_data
	     /logs
	     /output_data
	     /results
	         /tables
		 /plots
     /src
	...
     /test
	...

It is fairly simple to add additional layers under /modelspec, (eg subspecs, different datasets).
To do so, edit createSaveDirectories() in mSPEC.jl. However, note that at this time this will mess up the
path-getting functions currently in abstractdsgemodel.jl. (Note: consider moving those functions to mSPEC.jl)