Add catchGensysErrors flag to posterior and likelihood functions
Our goal is to catch GensysErrors during the posterior calculation and return a result that will be rejected by
the optimization. This approach involves passing an additional argument, catchGensysErrors, to posterior (which
propagates through to likelihood).

When catchGensysErrors is true, if a GensysError is thrown, we simply return -Inf.

posterior() sets CatchGensysErrors to true whenever mh=true. However, it can also be set to true by the caller, as
we do when reoptimize=true.