Matching Estimate Step Commit #1: Estimate step matches Matlab exactly, and is mat-file independent
   - Code contains large blocks of comments demonstrating how
     parameters/matrices were tested at various steps

   - Complete conversion of all test suite input to HDF5 files instead of mat files
       - runtests.jl tests
       - timing tests

   - Last commit containing mat files; they will be expunged in the commit to follow.

Given the same inputs, sigscale, and randvecs/randvals (acting as a
random number generator of limited size), the estimate step matches
the Matlab code exactly at every parameter draw in
Metropolis-Hastings. This commit contains commented-out code that
demonstrates how our Metropolis-Hastings was tested against the Matlab
version of the same, as well as code that read in mat files (which
have now been converted to HDF5). The commit immediately following
this one will remove the testing code and all code that refers to mat
files.