Some (incomplete) progress towards using the model's random number generator for testing.
Changes to the code:
-rand(d::DegenerateMvNormal, model, cc): can draw
from a DegenerateMvNormal using the model's random number generator

-Add metropolis_hastings.h5 for testing

Remarks:
- At the time of this commit, we get different testing results
when we include runtests.jl than we get when we call Pkg.test. The
discrepancy has been traced to the eigenvalue decomposition of the
hessian matrix that is computed in proposal_distribution, though we
don't know what is causing that decomposition to have different
results when the same code is being run between the `Pkg.test("DSGE")`
command and the `include("runtests.jl")` command.

- There's also a bunch of testing code/comments in this commit that
  will ultimately be removed.