Create separate test for eigenvalue decomposition to deal with nonsingular hessians
This mostly solves issue #1 in Eig.jl - namely, that `Pkg.test("DSGE")` and
`include("runtests.jl")` produced different results for
metropolis_hastings.jl.

To paraphrase @andreasnoack:

Eigenvectors of decomposition of singular matrices are not well
determined (because eigenvalues are effectively zero). Therefore,
small pertubations can change eigenvectors a lot. The number of
threads used by BLAS are set differently when you run `Pkg.test()`
than when you run `include("runtests.jl")`, and this slightly changes
the computation of the eigenvectors.

Because this is such a specific problem that could potentially affect
a lot of our end users, we add a new test for eigenvalue decomposition
to the test suite. In that test, we force the machine to only use 1
thread by calling `addprocs` before doing the eigenvalue decomposition.

The metropolis_hastings test has been refigured to read in the
covariance matrix of the proposal distribution directly. Therefore, if
there are problems with `eig.jl` and not with
`metropolis_hastings.jl`, we'll know that we haven't introduced any
bugs into metropolis_hastings.

TODO: Set appropriate thresholds for relative differences between expected and actual eig.jl output.