Add testing as a keyword argument for model constructor for Model990 and SmetsWouters types. TODO: implement this for m997 and m1002 types.
This keyword argument allows the model to be constructed in testing
mode, which suppresses warnings that result from having an empty
dataroot directory. Note that the <= syntax now does not add settings
to m.test_settings when m.testing=true. To add to m.test_settings, use
regular syntax for Dicts.