Simulation smoother runs
The simulation smoother runs (even with `n_anticipated_shocks(m) > 0`), but
`alpha_hat` and `eta_hat` do not equal the expected results read in from
[simulation_smoother_out.h5]. This is because it is extremely difficult to
replicate the same conditions in Matlab as in Julia: e.g. even with
`spec_990_715_2`, the TTT, RRR, and CCC matrices generated from the priors in
Matlab are not the same as those we get from solving the model in Julia with the
priors. From this, solving the discrete Lyapunov equation yields different P0
matrices, which eventually cause (I think) the smoothed states to differ
dramatically between Matlab and Julia.

As it stands, I think the simulation smoother is mostly correct, but we need a
better way of testing this code than on model 990.