fixed bug in wage Phillips curve
I think the term z_{t-1} is missing and coefficient for term z_t is incorrect in the Phillips curve for wages.
If we take the wage Phillips curve in Smets Wouters 2007 (page 6, equation 13) and replace pi_t with pi_t+z_t then we don't get the same equation as equation 14 in Chapter 2 of DSGE Model-Based Forecasting in Handbook of Economic Forecasting (which is also equation 8.106 derived incorrectly in the technical appendix).

The second bug is in the evolution of net worth in flexible economy. The term gamma_t is missing.